test/hyperbahn-client: move purge test into new file
