+++ test/index.js
@@ -79,6 +79,7 @@ require('./hyperbahn-client/forward-retry.js')(TestCluster);
 require('./hyperbahn-client/hyperbahn-down.js')(TestCluster);
 require('./hyperbahn-client/hyperbahn-times-out.js')(TestCluster);
 require('./hyperbahn-client/rate-limiter.js')(TestCluster);
+require('./hyperbahn-client/advertise-with-purge-interval.js')(TestCluster);
 
 require('./larch/errors.js');
 require('./larch/larch.js');
